refactor!: move interfaces and type aliases to types.ts
