+++ src/lib.ts
@@ -1,6 +1,7 @@
 export * from './_types';
+export * from './types';
 export * from './info';
 export * from './command';
-export { Progress, FFmpegProcess, spawn } from './process';
+export { spawn } from './process';
 export * from './probe';
 export { FFmpegError, FFprobeError } from './errors';
