+++ src/process.ts
@@ -1,105 +1,8 @@
-import {
-  ChildProcess, ChildProcessWithoutNullStreams,
-  spawn as spawnChildProcess
-} from 'child_process';
+import { ChildProcessWithoutNullStreams, spawn as spawnChildProcess } from 'child_process';
 import { createInterface as readlines } from 'readline';
+import { FFmpegProcess, Progress, SpawnOptions } from './types';
 import { pause, resume, write } from './utils';
 import { FFmpegError } from './errors';
-import type { SpawnOptions } from './command';
-
-/** @public */
-export interface Progress {
-  frames: number;
-  fps: number;
-  bitrate: number;
-  bytes: number;
-  time: number;
-  framesDuped: number;
-  framesDropped: number;
-  speed: number;
-}
-
-/** @public */
-export interface FFmpegProcess {
-  /**
-   * **UNSTABLE:** Deprecated, not for use in new projects.
-   *
-   * @deprecated Use `FFmpegProcess.unwrap().pid` instead.
-   *
-   * Returns the process identifier (PID) of the process.
-   */
-  readonly pid: number;
-  /** Command line arguments used to spawn the process. */
-  readonly args: readonly string[];
-  /** Path of the running ffmpeg executable. */
-  readonly ffmpegPath: string;
-  /**
-   * Returns an AsyncGenerator representing the real-time progress of the conversion.
-   * @example
-   * ```ts
-   * const process = await cmd.spawn();
-   * for await (const progress of process.progress()) {
-   *   console.log('Speed:', progress.speed);
-   * }
-   * ```
-   * Using NodeJS Streams:
-   * ```ts
-   * const process = await cmd.spawn();
-   * const progressStream = Readable.from(process.progress());
-   * progressStream.on('data', (progress) => {
-   *   console.log('Speed:', progress.speed);
-   * });
-   * ```
-   */
-  progress(): AsyncGenerator<Progress, void, void>;
-  /**
-   * Returns a Promise which resolves when the process exits, or rejects when the process exits with
-   * a non-zero status code. If the `ChildProcess` emits an `error` event, the Promise will be
-   * rejected with that error.
-   * @example
-   * ```ts
-   * const process = cmd.spawn();
-   * await process.complete();
-   * console.log('Conversion complete!');
-   * ```
-   * To handle errors:
-   * ```ts
-   * const process = await cmd.spawn();
-   * try {
-   *   await process.complete();
-   *   console.log('Conversion complete!');
-   * } catch (e) {
-   *   console.error('Conversion failed!', error);
-   * }
-   * ```
-   */
-  complete(): Promise<void>;
-  /**
-   * Aborts the conversion allowing FFmpeg to finish the generated files correctly.
-   * This waits for FFmpeg to exit but doesn't guarantee that FFmpeg will succeed,
-   * any possible errors should still be handled.
-   */
-  abort(): Promise<void>;
-  /** Returns the underlying NodeJS' ChildProcess instance. */
-  unwrap(): ChildProcess;
-  /**
-   * **UNSTABLE**: Deprecated, not for use in new projects.
-   *
-   * Sends a signal to the running process.
-   * See {@link https://nodejs.org/api/child_process.html#child_process_subprocess_kill_signal}
-   *
-   * @deprecated To terminate the conversion use {@link FFmpegProcess.abort}, to pause and resume
-   * the process use {@link FFmpegProcess.pause} or {@link FFmpegProcess.resume}. If you really
-   * have to send a signal to the process use `process.unwrap().kill(signal)`.
-   *
-   * @param signal - The signal to send.
-   */
-  kill(signal?: NodeJS.Signals | number): boolean;
-  /** Pauses the conversion, returns `true` if the operation succeeds or `false` if it fails. */
-  pause(): boolean;
-  /** Resumes the conversion, returns `true` if the operation succeeds or `false` if it fails. */
-  resume(): boolean;
-}
 
 /**
  * Start an FFmpeg process with the given arguments.
